[#36] Restructure route, refactor country-company ep
